third_party/WebKit/LayoutTests/fast/viewport/viewport-118-expected.txt

CONSOLE ERROR: line 3: The key "" is not recognized and ignored.
ALERT: viewport size 320x352 scale 1 with limits [1, 5] and userScalable true
